How to jpa

+++ src/main/java/com/green/gallery_jwt_jpa/account/AccountMapper.java
@@ -11,5 +11,5 @@ import java.util.List;
 public interface AccountMapper {
     int save(AccountJoinReq req);
     AccountLoginRes findByLoginId(AccountLoginReq req);
-    List<String> findAllRolesByMemberId(int memberId);
+    List<String> findAllRolesByMemberId(long memberId);
 }
