+++ src/main/java/com/green/gallery_jwt_jpa/cart/CartMapper.java
@@ -1,5 +1,6 @@
 package com.green.gallery_jwt_jpa.cart;
 
+
 import com.green.gallery_jwt_jpa.cart.model.CartDeleteReq;
 import com.green.gallery_jwt_jpa.cart.model.CartGetRes;
 import com.green.gallery_jwt_jpa.cart.model.CartPostReq;
@@ -10,7 +11,7 @@ import java.util.List;
 @Mapper
 public interface CartMapper {
     int save(CartPostReq req);
-    List<CartGetRes> findAllWithItemByMemberId(int memberId);
+    List<CartGetRes> findAllWithItemByMemberId(long memberId);
     int deleteByCartIdAndMemberId(CartDeleteReq req);
-    int deleteByMemberId(int memberId);
+    int deleteByMemberId(long memberId);
 }
