+++ src/main/java/com/green/gallery_jwt_jpa/cart/model/CartPostReq.java
@@ -6,6 +6,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class CartPostReq {
-    private int memberId;
-    private int itemId;
+    private long memberId;
+    private long itemId;
 }
