+++ src/main/java/com/green/gallery_jwt_jpa/config/model/JwtUser.java
@@ -8,7 +8,7 @@ import java.util.List;
 @Getter
 @AllArgsConstructor
 public class JwtUser {
-    private int signedUserId;
+    private long signedUserId;
     private List<String> roles; //인가 처리 때 사용
     /*
     role 이름은 ROLE_아무거나
