+++ src/main/java/com/green/gallery_jwt_jpa/order/OrderController.java
@@ -23,7 +23,7 @@ public class OrderController {
 
     @PostMapping
     public ResponseEntity<?> add(@AuthenticationPrincipal UserPrincipal userPrincipal, @RequestBody OrderPostReq req) {
-        int logginedMemberId = userPrincipal.getMemberId();
+        long logginedMemberId = userPrincipal.getMemberId();
         log.info("req: {}", req);
         int result = orderService.saveOrder(req, logginedMemberId);
         return ResponseEntity.ok(result);
@@ -32,14 +32,14 @@ public class OrderController {
 
     @GetMapping
     public ResponseEntity<?> findAllByMemberIdOrderByIdDesc(@AuthenticationPrincipal UserPrincipal userPrincipal) {
-        int logginedMemberId = userPrincipal.getMemberId();
+        long logginedMemberId = userPrincipal.getMemberId();
         List<OrderGetRes> result = orderService.findAllByMemberId(logginedMemberId);
         return ResponseEntity.ok(result);
     }
 
     @GetMapping("/{orderId}")
     public ResponseEntity<?> findDetail(@AuthenticationPrincipal UserPrincipal userPrincipal, @PathVariable int orderId) {
-        int logginedMemberId = userPrincipal.getMemberId();
+        long logginedMemberId = userPrincipal.getMemberId();
         OrderDetailGetReq req = new OrderDetailGetReq();
         req.setOrderId(orderId);
         req.setMemberId(logginedMemberId);
